+++ braintree/transaction.py
@@ -158,6 +158,7 @@ class Transaction(Resource):
         SettlementConfirmed    = "settlement_confirmed"
         SettlementDeclined     = "settlement_declined"
         SettlementFailed       = "settlement_failed"
+        SettlementPending      = "settlement_pending"
         Settling               = "settling"
         SubmittedForSettlement = "submitted_for_settlement"
         Voided                 = "voided"
@@ -442,7 +443,8 @@ class Transaction(Resource):
                 ]
             },
             {"custom_fields": ["__any_key__"]},
-            {"descriptor": ["name", "phone", "url"]}
+            {"descriptor": ["name", "phone", "url"]},
+            {"paypal_account": ["payee_email"]}
         ]
 
     def __init__(self, gateway, attributes):