+++ tests/test_helper.py
@@ -92,6 +92,10 @@ class TestHelper(object):
         return Configuration.gateway().testing.settlement_decline_transaction(transaction_id)
 
     @staticmethod
+    def settlement_pending_transaction(transaction_id):
+        return Configuration.gateway().testing.settlement_pending_transaction(transaction_id)
+
+    @staticmethod
     def simulate_tr_form_post(post_params, url=TransparentRedirect.url()):
         form_data = urlencode(post_params)
         conn = HTTPConnection(Configuration.environment.server_and_port)
@@ -145,10 +149,9 @@ class TestHelper(object):
         return False
 
     @staticmethod
-    def now_in_eastern():
-        now  = datetime.utcnow()
-        offset  = timedelta(hours=5)
-        return (now - offset).strftime("%Y-%m-%d")
+    def now_minus_offset(offset):
+        now = datetime.utcnow()
+        return (now - timedelta(hours=offset)).strftime("%Y-%m-%d")
 
     @staticmethod
     def unique(list):