+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.4.0
+
+* Added paypal specific fields to transaction calls
+* Added SettlementPending, SettlementDeclined transaction statuses
+
 ## 3.3.0
 
 * Add Descriptor url support